Prepare shared campaign writes

+++ src/components/SessionItem/SessionItem.js
@@ -30,67 +30,78 @@ class SessionItem extends Component {
 	}
 
 	// Triggers when deleting a session
-	deleteSession() {
+	deleteSession(attempted) {
     
 		// Delete recaps from tags locally and from recaps collection on Firestore
 		let tags = this.props.tags;
-		let recaps = this.props.sessions[this.props.sessionID].recaps
+		let recaps = this.props.sessions[this.props.sessionID].recaps;
 		let campaign = this.props.campaign;
 
-		for(let recap in recaps) {
+		if(!attempted) {
+			for(let recap in recaps) {
 
-			// Delete from tags locally
-			recaps[recap].tags.forEach(tag => {
-				delete tags[tag].recaps[recap];
-			});
-
-			// Delete on Firestore
-			this.props.campaignRef.collection("recaps").doc(recap).delete()
-			.then(() => {
-				console.log("Document successfully deleted!");
-			}).catch((error) => {
-				console.log("Error deleting document:", error);
-			});
+				// Delete from tags locally
+				recaps[recap].tags.forEach(tag => {
+					delete tags[tag].recaps[recap];
+				});
+				
+				// Delete on Firestore
+				this.props.campaignRef.collection("recaps").doc(recap).delete()
+				.then(() => {
+					console.log("Recap successfully deleted!");
+				}).catch((error) => {
+					console.log("Error deleting recap:", error);
+				});
+			}
+			this.props.handleTags(tags);
 		}
-		this.props.handleTags(tags);
-
-		// Delete session recaps locally
-		let sessions = this.props.sessions;
-		delete sessions[this.props.sessionID];
-		this.props.handleSessions(sessions);		
-
-		// Delete session info locally
-		delete campaign.sessions[this.props.sessionID];
 
 		// Delete session from the session order list
-		let sessionIndex = campaign.sessionOrder.indexOf(this.props.sessionID);
-		if (sessionIndex !== -1) campaign.sessionOrder.splice(sessionIndex, 1);
-		this.props.handleCampaign(campaign);
+		let sessionOrder = [...campaign.sessionOrder];
+		let sessionIndex = sessionOrder.indexOf(this.props.sessionID);
+		if (sessionIndex !== -1) sessionOrder.splice(sessionIndex, 1);
 
 		// Delete session info on Firestore
 		this.props.campaignRef.update({
 			operation: "session-delete",
 			["sessions." + this.props.sessionID]: firebase.firestore.FieldValue.delete(),
-			sessionOrder: campaign.sessionOrder,
+			sessionOrder: sessionOrder,
 			selectedSession: this.props.sessionID,
 		}).then(() => {
-			console.log("Document successfully deleted!");
+			console.log("Session successfully deleted!");
+
+			// Delete session recaps locally
+			let sessions = this.props.sessions;
+			delete sessions[this.props.sessionID];
+			this.props.handleSessions(sessions);		
+
+			// Delete session info locally
+			delete campaign.sessions[this.props.sessionID];
+			campaign.sessionOrder = sessionOrder;
+			this.props.handleCampaign(campaign);
 
 			// Update selected session to the last session chronologically
-			let latestSession = campaign.sessionOrder[campaign.sessionOrder.length - 1] ? 
-				campaign.sessionOrder[campaign.sessionOrder.length - 1] : "";
+			let latestSession = sessionOrder[0] ? sessionOrder[0] : "";
 			this.props.handleSelectedSession(latestSession);
 
 			this.props.campaignRef.update({
 				operation: "selected-session-update",
 				selectedSession: latestSession,
 			}).then(() => {
-				console.log("Document successfully deleted!");
+				console.log("Selected session successfully updated");
 			}).catch((error) => {
-				console.log("Error deleting document:", error);
+				console.log("Error updating selected session:", error);
 			});
 		}).catch((error) => {
-			console.log("Error deleting document:", error);
+			console.log("Error deleting session:", error);
+			if(attempted) {
+				this.props.handleError(error, "Could not delete session")
+			} else {
+				console.log("Reading and reattempting");
+				this.props.loadCampaign(
+					() => {this.deleteSession(true)}
+				);
+			}
 		});
 	}
 
@@ -130,7 +141,7 @@ class SessionItem extends Component {
 										this.state.showIcons ?
 										<ItemMenu
 											edit = {() => this.setState({ showSessionInfo: true})}
-											delete = {this.deleteSession}
+											delete = {() => this.deleteSession(false)}
 											deleteText = {deleteText}
 										/> :
 										<div></div>
@@ -151,10 +162,12 @@ class SessionItem extends Component {
 					campaign = {this.props.campaign}
 					handleSessions = {this.props.handleSessions}
 					handleCampaign = {this.props.handleCampaign}
+					handleError = {this.props.handleError}
 					campaignRef = {this.props.campaignRef}
 					edit = {true}
 					session = {session}
 					sessionID = {this.props.sessionID}
+					loadCampaign = {this.props.loadCampaign}
 				/>
 			</>
 		);
